service/requirements.txt
efficientnet
tensorflow
matplotlib
aiofiles
Pillow
keras